change data provider context
