+++ AElf.Kernel/IAccountDataContext.cs
@@ -1,0 +1,7 @@
+﻿namespace AElf.Kernel
+{
+    public interface IAccountDataContext
+    {
+        ulong IncreasementId { get; set; }
+    }
+}
